refactor: remove unnecessary file from demo
