+++ app/src/main/java/com/codesoom/todo/App.java
@@ -1,14 +1,10 @@
-package com.codesoom.assignment;
+package com.codesoom.todo;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
 public class App {
-    public String getGreeting() {
-        return "Hello, world!";
-    }
-
     public static void main(String[] args) {
         SpringApplication.run(App.class, args);
     }
